SPRING: export META-INF package from u.a.d.org.springframework plugin
When using Eclipse Gemini Blueprint, an OsgiBundleXmlApplicationContext
is automatically created for bundles/plugins containing Spring XML files
in a standard location. To resolve the schemas used in the Spring
XML files, a BundleDelegatingClassLoader backed by the plugin is asked
to find resources named "META-INF/spring.schemas". The schema mappings
are not found unless u.a.d.org.springframework exports META-INF.