Removes incorrect jre.compilation.profile from bundles
jre.compilation.profile was set to Java 1.6 or Java 1.7 in several
bundles. This will fails any Tycho build as the
Bundle-RequiredExecutionEnvironment (BREE) in the MANIFEST.MF of the
affected bundles are already configured to use Java 8 and Tycho prefers
jre.compilation.profile over the BREE.

By removing it, we ensure that the BREE from the manifest is used bv
Tycho.

Change-Id: Ic9b50c77a0aaf9b5626930b4c17fdab5875990c6
Signed-off-by: Lars Vogel <Lars.Vogel@vogella.com>