DAQ-3104 Create uk.ac.diamond.org.springframework.spring-test
This bundle is to wrap spring-test into an OSGi bundle
its removal should be considered if we look at moving Spring to the
target platform (at this point it will probably be replaced with
org.apache.servicemix.bundles.spring-test) see DAQ-325.

There is no runtime guarantee that all the classes here
will work as Dynamic import package has been used
to satisfy its dependencies.

It is in a separate bundle to the main Diamond Spring bundle to prevent
it from being present in the runtime applications.

It can be used in test plugins by adding a Require-Bundle dependency to:
uk.ac.diamond.org.springframework.spring-test or the appropriate
Import-Package.

Change-Id: I3267c4787d951c10a43a471033b335848e80a9ad