DAQ-3104 Build/execution fixes for spring-test
Using purely DynamicImport-Pacakge did not seem to be sufficient to
enable the bundle to load things like BlockJUnit4ClassRunner. Instead
provide a proper manifest generated by the maven-bundle-plugin.

The package also contains optional packages which we don't have in the
target platform so the pom is required to ignore these from the "build".

Change-Id: I2c69dd93e4805b2a5318613af3e078a579ca7e80